@@ -2,17 +2,15 @@
   "name": "fill-the-fridge",
   "version": "1.0.0",
   "description": "A web service, which helps you to make a dish from ingredients, which you have in your fridge.\r If you don't have neccesary inridients, this service will find several options for stores and show you the best price for the unit. You will be able to select the nearest store and the store with lowest price.",
-  "main": "index.js",
   "scripts": {
-    "debug": "cross-env DEBUG=index open-cli http://localhost:3000 & nodemon index.js",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
     "test": "npm run lint",
-    "prettier": "prettier --config .prettierrc.json --write index.js",
-    "lint": "eslint ./server/index.js & eslint ./client/**/*.js",
-    "lint:fix": "eslint --fix ./server/index.js & eslint --fix ./client/**/*.js",
+    "lint:fix": "npm run lint -- --fix",
     "precommit": "npm run lint:fix",
-    "prepush": "npm run test",
-    "start": "node ./server/index.js",
-    "dev": "nodemon ./server/index.js"
+    "prepush": "npm run test"
   },
   "repository": {
     "type": "git",
@@ -25,46 +23,21 @@
     "url": "https://github.com/MetaMmodern/Fill-the-Fridge/issues"
   },
   "homepage": "https://github.com/MetaMmodern/Fill-the-Fridge#readme",
-  "devDependencies": {
-    "@babel/cli": "^7.10.1",
-    "@babel/core": "^7.10.2",
-    "@babel/preset-env": "^7.10.2",
-    "babel-eslint": "^10.1.0",
-    "chalk": "^4.0.0",
-    "cross-env": "^7.0.2",
-    "debug": "^4.1.1",
-    "eslint": "^7.2.0",
-    "eslint-config-airbnb": "^18.1.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.21.2",
-    "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "husky": "^3.1.0",
-    "nodemon": "^2.0.16",
-    "open-cli": "^6.0.1",
-    "pre-commit": "^1.2.2",
-    "prettier": "^1.19.1"
-  },
   "dependencies": {
-    "@babel/polyfill": "^7.10.1",
-    "cheerio": "^1.0.0-rc.3",
-    "fetch": "^1.1.0",
-    "iconv-lite": "^0.5.1",
-    "koa": "^2.11.0",
-    "koa-body": "^4.1.1",
-    "koa-bodyparser": "^4.3.0",
-    "koa-ejs": "^4.3.0",
-    "koa-morgan": "^1.0.1",
-    "koa-router": "^8.0.8",
-    "koa-static": "^5.0.0",
-    "morgan": "^1.10.0",
-    "needle": "^2.5.0",
-    "node-fetch": "^2.6.1",
-    "request": "^2.88.2",
-    "rotating-file-stream": "^2.0.2",
-    "superagent": "^5.2.2"
+    "bootstrap": "^5.1.3",
+    "cheerio": "^1.0.0-rc.10",
+    "needle": "^3.1.0",
+    "next": "12.1.6",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "sass": "^1.51.0"
+  },
+  "devDependencies": {
+    "@types/node": "17.0.33",
+    "@types/react": "18.0.9",
+    "@types/react-dom": "18.0.4",
+    "eslint": "8.15.0",
+    "eslint-config-next": "12.1.6",
+    "typescript": "4.6.4"
   }
 }
